glympse: Add internal temperature reading
This adds the P7MU temperature reading to the Glympse string. It is the most
sensible internal temperature sensor I could find.

@AnthonyH has tested its' limits and found the Chuck starts flashing red at
103 degrees and shuts down at 111.

To make room for this new value in the Glympse string, I have removed the
prefixes for each value. To make the battery reading read more intuitive I
also swapped the positions of battery volt and battery percent.